backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
groq
python-dotenv
python-multipart
PyPDF2>=3.0
python-docx>=1.0
upstash-redis
sqlmodel
psycopg2-binary
slowapi
cryptography
